// frontend/src/shared/api/generated/model/getCollectionItemsCollectionItemIdPlaySessions200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Boardgames tracker API
 * API для проекта Boardgames tracker
 * OpenAPI spec version: 1.0.0
 */

export type GetCollectionItemsCollectionItemIdPlaySessions200DataItem = {
  readonly id: number;
  readonly collectionItemId: number;
  dateStart: string;
  /** @nullable */
  dateEnd?: string | null;
  /**
   * @maxLength 2000
   * @nullable
   */
  comment?: string | null;
  readonly createdAt: string;
  readonly updatedAt: string;
};