// frontend/src/shared/api/generated/model/postCollectionItemsCollectionItemIdPlaySessionsBody.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Boardgames tracker API
 * API для проекта Boardgames tracker
 * OpenAPI spec version: 1.0.0
 */

export type PostCollectionItemsCollectionItemIdPlaySessionsBody = {
  dateStart: string;
  /** @nullable */
  dateEnd?: string | null;
  /**
   * @maxLength 2000
   * @nullable
   */
  comment?: string | null;
};